07.01.24 14:19
Полностью готова система удаления аккаунта. Нужно запомнить: ExecuteNonQuery - для сохранения и отправки изменённых данных в бд.

+++ GiveMePaw/Account.cs
@@ -26,7 +26,7 @@ namespace GiveMePaw
             MySqlDataAdapter adapter = new MySqlDataAdapter();
             DB db = new DB();
             db.openConnection();
-            MySqlCommand command = new MySqlCommand("SELECT second_name FROM users WHERE email = @email", db.getConnection());
+            MySqlCommand command = new MySqlCommand("SELECT * FROM users WHERE email = @email", db.getConnection());
             MySqlParameter n1 = new MySqlParameter("@email", SignIn.user_email);
             command.Parameters.Add(n1);
             MySqlDataReader Reader = command.ExecuteReader();
@@ -100,101 +100,124 @@ namespace GiveMePaw
 
         private void cancel_button_account_Click(object sender, EventArgs e)
         {
-            DataTable table = new DataTable();
-            MySqlDataAdapter adapter = new MySqlDataAdapter();
-            DB db = new DB();
-            db.openConnection();
-            MySqlCommand command = new MySqlCommand("SELECT second_name FROM users WHERE email = @email", db.getConnection());
-            MySqlParameter n1 = new MySqlParameter("@email", SignIn.user_email);
-            command.Parameters.Add(n1);
-            MySqlDataReader Reader = command.ExecuteReader();
-
-            object surname = "";
-            object name = "";
-            object last_name = "";
-            object phone = "";
-            object password = "";
-            if (Reader.HasRows)
+            if (MessageBox.Show("Вы уверены, что хотите отменить изменения?", "Информация", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Information) == DialogResult.Yes)
             {
-                while (Reader.Read())
+                try
                 {
-                    surname = Reader.GetValue(2);
-                    name = Reader.GetValue(1);
-                    last_name = Reader.GetValue(3);
-                    phone = Reader.GetValue(5);
-                    password = Reader.GetValue(6);
+                    DataTable table = new DataTable();
+                    MySqlDataAdapter adapter = new MySqlDataAdapter();
+                    DB db = new DB();
+                    db.openConnection();
+                    MySqlCommand command = new MySqlCommand("SELECT * FROM users WHERE email = @email", db.getConnection());
+                    MySqlParameter n1 = new MySqlParameter("@email", SignIn.user_email);
+                    command.Parameters.Add(n1);
+                    MySqlDataReader Reader = command.ExecuteReader();
+
+                    object surname = "";
+                    object name = "";
+                    object last_name = "";
+                    object phone = "";
+                    object password = "";
+                    if (Reader.HasRows)
+                    {
+                        while (Reader.Read())
+                        {
+                            surname = Reader.GetValue(2);
+                            name = Reader.GetValue(1);
+                            last_name = Reader.GetValue(3);
+                            phone = Reader.GetValue(5);
+                            password = Reader.GetValue(6);
+                        }
+                    }
+                    db.closeConnection();
+
+                    surname_account.Text = Convert.ToString(surname);
+                    name_account.Text = Convert.ToString(name);
+                    last_name_account.Text = Convert.ToString(last_name);
+                    email_account.Text = SignIn.user_email;
+                    phone_account.Text = Convert.ToString(phone);
+                    password_account.Text = Convert.ToString(password);
+
+                    surname_account.ReadOnly = true;
+                    name_account.ReadOnly = true;
+                    last_name_account.ReadOnly = true;
+                    password_account.ReadOnly = true;
+
+                    surname_account.ForeColor = Color.Gray;
+                    name_account.ForeColor = Color.Gray;
+                    last_name_account.ForeColor = Color.Gray;
+                    email_account.ForeColor = Color.Gray;
+                    phone_account.ForeColor = Color.Gray;
+                    password_account.ForeColor = Color.Gray;
+
+                    update_button_account.Visible = true;
+                    save_button_account.Visible = false;
+                    cancel_button_account.Visible = false;
+                    delete_button_account.Visible = false;
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка!" + ex);
                 }
             }
-            db.closeConnection();
-
-            surname_account.Text = Convert.ToString(surname);
-            name_account.Text = Convert.ToString(name);
-            last_name_account.Text = Convert.ToString(last_name);
-            email_account.Text = SignIn.user_email;
-            phone_account.Text = Convert.ToString(phone);
-            password_account.Text = Convert.ToString(password);
-
-            surname_account.ReadOnly = true;
-            name_account.ReadOnly = true;
-            last_name_account.ReadOnly = true;
-            password_account.ReadOnly = true;
-
-            surname_account.ForeColor = Color.Gray;
-            name_account.ForeColor = Color.Gray;
-            last_name_account.ForeColor = Color.Gray;
-            email_account.ForeColor = Color.Gray;
-            phone_account.ForeColor = Color.Gray;
-            password_account.ForeColor = Color.Gray;
-
-            update_button_account.Visible = true;
-            save_button_account.Visible = false;
-            cancel_button_account.Visible = false;
-            delete_button_account.Visible = false;
+            
+            
         }
 
         private void save_button_account_Click(object sender, EventArgs e)
         {
+            if (MessageBox.Show("Вы уверены, что хотите сохранить изменения?", "Информация", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Information) == DialogResult.Yes)
+            {
+                try
+                {
+                    DataTable table = new DataTable();
+                    MySqlDataAdapter adapter = new MySqlDataAdapter();
+                    DB db = new DB();
+                    db.openConnection();
 
-            DataTable table = new DataTable();
-            MySqlDataAdapter adapter = new MySqlDataAdapter();
-            DB db = new DB();
-            db.openConnection();
+                    MySqlCommand command = new MySqlCommand("UPDATE users SET name=@name, second_name=@surname, patronomic=@last_name, email=@new_email, phone_number=@phone, password=@password WHERE email = @old_email", db.getConnection());
+                    MySqlParameter n1 = new MySqlParameter("@old_email", SignIn.user_email);
+                    command.Parameters.Add(n1);
+                    MySqlParameter n2 = new MySqlParameter("@surname", surname_account.Text);
+                    command.Parameters.Add(n2);
+                    MySqlParameter n3 = new MySqlParameter("@name", name_account.Text);
+                    command.Parameters.Add(n3);
+                    MySqlParameter n4 = new MySqlParameter("@new_email", email_account.Text);
+                    command.Parameters.Add(n4);
+                    MySqlParameter n5 = new MySqlParameter("@phone", phone_account.Text);
+                    command.Parameters.Add(n5);
+                    MySqlParameter n6 = new MySqlParameter("@password", password_account.Text);
+                    command.Parameters.Add(n6);
+                    MySqlParameter n7 = new MySqlParameter("@last_name", last_name_account.Text);
+                    command.Parameters.Add(n7);
+                    command.ExecuteNonQuery();
+                    db.closeConnection();
 
-            MySqlCommand command = new MySqlCommand("UPDATE users SET name=@name, second_name=@surname, patronomic=@last_name, email=@new_email, phone_number=@phone, password=@password WHERE email = @old_email", db.getConnection());
-            MySqlParameter n1 = new MySqlParameter("@old_email", SignIn.user_email);
-            command.Parameters.Add(n1);
-            MySqlParameter n2 = new MySqlParameter("@surname", surname_account.Text);
-            command.Parameters.Add(n2);
-            MySqlParameter n3 = new MySqlParameter("@name", name_account.Text);
-            command.Parameters.Add(n3);
-            MySqlParameter n4 = new MySqlParameter("@new_email", email_account.Text);
-            command.Parameters.Add(n4);
-            MySqlParameter n5 = new MySqlParameter("@phone", phone_account.Text);
-            command.Parameters.Add(n5);
-            MySqlParameter n6 = new MySqlParameter("@password", password_account.Text);
-            command.Parameters.Add(n6);
-            MySqlParameter n7 = new MySqlParameter("@last_name", last_name_account.Text);
-            command.Parameters.Add(n7);
-            command.ExecuteNonQuery();
-            db.closeConnection();
+                    surname_account.ReadOnly = true;
+                    name_account.ReadOnly = true;
+                    last_name_account.ReadOnly = true;
+                    password_account.ReadOnly = true;
 
-            surname_account.ReadOnly = true;
-            name_account.ReadOnly = true;
-            last_name_account.ReadOnly = true;
-            password_account.ReadOnly = true;
+                    surname_account.ForeColor = Color.Gray;
+                    name_account.ForeColor = Color.Gray;
+                    last_name_account.ForeColor = Color.Gray;
+                    email_account.ForeColor = Color.Gray;
+                    phone_account.ForeColor = Color.Gray;
+                    password_account.ForeColor = Color.Gray;
 
-            surname_account.ForeColor = Color.Gray;
-            name_account.ForeColor = Color.Gray;
-            last_name_account.ForeColor = Color.Gray;
-            email_account.ForeColor = Color.Gray;
-            phone_account.ForeColor = Color.Gray;
-            password_account.ForeColor = Color.Gray;
+
+                    update_button_account.Visible = true;
+                    save_button_account.Visible = false;
+                    cancel_button_account.Visible = false;
+                    delete_button_account.Visible = false;
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show("Ошибка!" + ex);
+                }
+            }
             
-
-            update_button_account.Visible = true;
-            save_button_account.Visible = false;
-            cancel_button_account.Visible = false;
-            delete_button_account.Visible = false;
+            
         }
 
 
@@ -259,19 +282,22 @@ namespace GiveMePaw
 
         private void exit_button_account_Click(object sender, EventArgs e)
         {
-            try
+            if (MessageBox.Show("Вы уверены, что хотите выйти из аккаунта?", "Информация", MessageBoxButtons.YesNoCancel, MessageBoxIcon.Information) == DialogResult.Yes)
             {
-                File.Delete("checkSignIn.txt");
-                Account.ActiveForm.Hide();
-                SignIn NewForm = new SignIn();
-                NewForm.ShowDialog();
-                Close();
-                
+                try
+                {
+                    File.Delete("checkSignIn.txt");
+                    Account.ActiveForm.Hide();
+                    SignIn NewForm = new SignIn();
+                    NewForm.ShowDialog();
+                    Close();
+                }
+                catch
+                {
+                    MessageBox.Show("Ошибка!");
+                }
             }
-            catch
-            {
-                MessageBox.Show("Ошибка!");
-            }
+            
         }
 
         private void delete_button_account_Click(object sender, EventArgs e)
@@ -287,6 +313,7 @@ namespace GiveMePaw
                     MySqlCommand command = new MySqlCommand("DELETE FROM users WHERE email = @email", db.getConnection());
                     MySqlParameter n1 = new MySqlParameter("@email", SignIn.user_email);
                     command.Parameters.Add(n1);
+                    command.ExecuteNonQuery();
                     db.closeConnection();
                     File.Delete("checkSignIn.txt");
                     Account.ActiveForm.Hide();
